Moved logic to DrsService
