+++ service/src/test/java/org/databiosphere/workspacedataservice/common/DataPlaneTestBase.java
@@ -36,7 +36,7 @@ import org.springframework.test.context.TestPropertySource;
       "rest.retry.backoff.delay=3",
       // Rawls url must be valid, else context initialization (Spring startup) will fail
       "rawlsUrl=https://localhost/",
-      "drshuburl=https://localhost/"
+      "drshubUrl=https://localhost/"
     })
 @ExtendWith(ConfigurationExceptionDetector.class)
 public abstract class DataPlaneTestBase {
